+++ packages/analysis/src/index.js
@@ -6,6 +6,3 @@ export { default as InSightAnalysis } from './InSightAnalysis.js';
 export { default as ExcavateAnalysis } from './ExcavateAnalysis.js';
 export { default as CrossCutAnalysis } from './CrossCutAnalysis.js';
 export { default as HeightLimitAnalysis } from './HeightLimitAnalysis.js';
-export { default as Height3DTool } from './Height3DTool.js';
-export { default as Area3DTool } from './Area3DTool.js';
-export { default as Distance3DTool } from './Distance3DTool.js';