+++ packages/layer-gltf/rollup.config.js
@@ -155,7 +155,7 @@ module.exports = [
         output: {
             globals: {
                 'maptalks': 'maptalks',
-                '@maptalks/gl': 'maptalksgl'
+                '@maptalks/gl': 'maptalks'
             },
             banner,
             outro,
@@ -198,7 +198,7 @@ if (production) {
             output: {
                 globals: {
                     'maptalks': 'maptalks',
-                    '@maptalks/gl': 'maptalksgl'
+                    '@maptalks/gl': 'maptalks'
                 },
                 banner,
                 outro,