+++ packages/vt/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @maptalks/vt
 
+## 0.103.2
+
+### Patch Changes
+
+- fix bundle export and analysis's import of reshader.gl
+- Updated dependencies
+  - @maptalks/gl@0.104.1
+  - @maptalks/vt-plugin@0.99.1
+
+## 0.103.1
+
+### Patch Changes
+
+- remove log
+
 ## 0.103.0
 
 ### Minor Changes